rename folder src/wallets to src/connectors
