+++ CHANGELOG.md
@@ -82,7 +82,7 @@ Special thanks to @re2005 and @Mulander-J for their contributions!
 #### Features
 
 - New logo from issue #24 (Special thank @ramuta for his contribution!)
-- Add [Metamask switch chain API](https://github.com/chnejohnson/vue-dapp/blob/062acd4092b8f6a2a67af750f2569294c90c4f7b/src/wallets/metamask.ts#L47) and [network constants](https://github.com/chnejohnson/vue-dapp/blob/main/src/constants/chainId.ts)
+- Add [Metamask switch chain API](https://github.com/chnejohnson/vue-dapp/blob/062acd4092b8f6a2a67af750f2569294c90c4f7b/src/connectors/metamask.ts#L47) and [network constants](https://github.com/chnejohnson/vue-dapp/blob/main/src/constants/chainId.ts)
 - Update documentation to latest.
 - Support connecting to Arbitrum, RinkArby, and Polygon with MetaMask
