+++ src/connectors/errors.ts
