+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './constants'
 export * from './composables'
-export * from './wallets'
+export * from './connectors'
 export * from './utils'
 export * from './plugin'
